+++ lib/indexes/common.d.ts
@@ -4,9 +4,15 @@ export declare type TargetProperty = string;
 export declare type IndexOptions<T extends IndexableObj> = {
     targetProperties: Properties;
 };
-export declare type Captor<T extends IndexableObj> = (obj: T) => T;
+export declare type IndexedObj<T extends IndexableObj> = Omit<T, 'deleteFromIndex' | 'getTarget' | 'isProxy'> & {
+    deleteFromIndex: () => void;
+    getTarget: () => T;
+    isProxy: true;
+};
+export declare type Captor<T extends IndexableObj> = (obj: T) => IndexedObj<T>;
 export declare type Updater<T extends IndexableObj> = (obj: T, propName: string, newValue: T[string]) => void;
 export declare type Ingestor<T extends IndexableObj> = (obj: T) => void;
+export declare type Deleter<T extends IndexableObj> = (obj: T) => void;
 export declare class IndexError extends Error {
 }
 export declare class MissingIndex extends IndexError {