+++ lib/indexes/noop.d.ts
@@ -1,0 +1,3 @@
+import { Captor, IndexableObj } from './common';
+export declare function createNoopCaptor<T extends IndexableObj>(): Captor<T>;
+//# sourceMappingURL=noop.d.ts.map