+++ src/indexes/common.ts
@@ -6,14 +6,21 @@ export type TargetProperty = string;
 export type IndexOptions<T extends IndexableObj> = {
   targetProperties: Properties;
 };
+export type IndexedObj<T extends IndexableObj> = T &
+  {
+    deleteFromIndex: () => void;
+    getTarget: () => T;
+    isProxy: true;
+  }
 
-export type Captor<T extends IndexableObj> = (obj: T) => T;
+export type Captor<T extends IndexableObj> = (obj: T) => IndexedObj<T>;
 export type Updater<T extends IndexableObj> = (
   obj: T,
   propName: string,
   newValue: T[string]
 ) => void;
 export type Ingestor<T extends IndexableObj> = (obj: T) => void;
+export type Deleter<T extends IndexableObj> = (obj: T) => void;
 
 export class IndexError extends Error {}
 export class MissingIndex extends IndexError {}